src/spider/code/requirements.txt
auto_mix_prep==0.2.0
elasticsearch==8.17.0
fastapi==0.115.6
langchain==0.3.13
langchain_community==0.3.13
langchain_openai==0.2.14
langgraph==0.2.60
llama_index==0.12.9
loguru==0.6.0
mysql_connector_repackaged==0.3.1
pydantic==2.10.4
Requests==2.32.3
uvicorn==0.34.0
